Enables dynamic theme compilation
Introduces a new command to compile theme CSS from the database,
allowing for dynamic theme customization based on database
settings and Tailwind CSS.

Replaces the previous file-based theme generation with a dynamic
compilation process using `@tailwindcss/cli`, storing source CSS
in a temporary file and outputting the compiled CSS to the public
directory.

Removes old theme files and introduces a script to generate a theme
from CMS content during development, using PostCSS and Tailwind.

Updates the `package.json` to include `@tailwindcss/cli` as a
dependency and update the tailwindcss version, and includes
`dynamic-theme.css` in the blade template to allow themes to be
rendered dynamically.