Enhances theme customization with radius options and auto compile on Update
Adds radius size variables to the theme and updates the theme compilation process.

Includes a temporary fix to replace potentially incorrect radius names in the database with the correct tailwind class names.

Introduces a background job to compile theme CSS to improve the responsiveness of the theme customizer when saving.